Move Summary class
to avoid an import cycle in the next changes
